Make tests runnable
