+++ build.gradle.kts
@@ -63,6 +63,8 @@ recipeDependencies {
     parserClasspath("org.springframework.boot:spring-boot-actuator:2.5.+")
     parserClasspath("org.springframework.boot:spring-boot-test:2.+")
 
+    parserClasspath("org.springframework:spring-jdbc:4.1.+")
+
     parserClasspath("org.springframework:spring-beans:4.+")
     parserClasspath("org.springframework:spring-beans:5.+")
     parserClasspath("org.springframework:spring-beans:6.+")
