+++ src/main/java/org/openrewrite/java/spring/data/JdbcTemplateQueryForLongMigration.java
@@ -1,20 +1,19 @@
 /*
- * Copyright (c) The original author or authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * Copyright 2024 the original author or authors.
+ * <p>
+ * Licensed under the Moderne Source Available License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * https://docs.moderne.io/licensing/moderne-source-available-license
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
-
-package org.openrewrite.java.spring;
+package org.openrewrite.java.spring.data;
 
 import org.openrewrite.ExecutionContext;
 import org.openrewrite.Recipe;
@@ -26,12 +25,12 @@ import org.openrewrite.java.tree.J;
 public class JdbcTemplateQueryForLongMigration extends Recipe {
     @Override
     public String getDisplayName() {
-        return "Convert JdbcTemplate.queryForLong() to queryForObject()";
+        return "Convert `JdbcTemplate.queryForLong()` to `queryForObject()`";
     }
 
     @Override
     public String getDescription() {
-        return "- [Spring Framework Migration] Replaces calls to JdbcTemplate.queryForLong() with queryForObject(String, Class, Object...).";
+        return "Replaces calls to `JdbcTemplate.queryForLong()` with `queryForObject(String, Class, Object...)`.";
     }
 
     @Override
